+++ packages/orq-rc/src/models/components/piiredactionpluginauto.ts
@@ -45,7 +45,6 @@ export const Entities = {
   DateTime: "DATE_TIME",
   DeviceId: "DEVICE_ID",
   EducationLevel: "EDUCATION_LEVEL",
-  Email: "EMAIL",
   EmailAddress: "EMAIL_ADDRESS",
   EmploymentStatus: "EMPLOYMENT_STATUS",
   Gender: "GENDER",