+++ packages/orq-rc/src/models/components/piiredactionpluginen.ts
@@ -58,7 +58,6 @@ export const PIIRedactionPluginEnEntities = {
   DateTime: "DATE_TIME",
   DeviceId: "DEVICE_ID",
   EducationLevel: "EDUCATION_LEVEL",
-  Email: "EMAIL",
   EmailAddress: "EMAIL_ADDRESS",
   EmploymentStatus: "EMPLOYMENT_STATUS",
   Gender: "GENDER",