+++ packages/orq-rc/RELEASES.md
@@ -10429,3 +10429,13 @@ Based on:
 - [typescript v4.11.0-rc.65] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.65] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.65 - packages/orq-rc
+
+## 2026-07-01 12:28:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.66] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.66] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.66 - packages/orq-rc